requirements.txt
python-telegram-bot[job-queue]==21.7
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
